+++ examples/button.js
@@ -1,38 +1,38 @@
 // simple UI
 
-kaboom({
-	scale: 2,
-});
+kaboom();
 
 function addButton(txt, p, f) {
 
-	const bg = add([
-		pos(p),
-		rect(60, 30),
-		area(),
-		origin("center"),
-		color(1, 1, 1),
-	]);
-
-	add([
+	const btn = add([
 		text(txt, 8),
 		pos(p),
+		area(),
+		scale(1),
 		origin("center"),
-		color(0, 0, 0),
 	]);
 
-	bg.action(() => {
-		if (bg.isHovered()) {
-			bg.color = rgb(0.8, 0.8, 0.8);
+	btn.action(() => {
+		if (btn.isHovered()) {
+			const t = time() * 10;
+			btn.color = rgb(
+				wave(0, 1, t),
+				wave(0, 1, t + 2),
+				wave(0, 1, t + 4),
+			);
+			btn.scale = vec2(1.2);
+			cursor("pointer");
 			if (mouseIsClicked()) {
 				f();
 			}
 		} else {
-			bg.color = rgb(1, 1, 1);
+			cursor("default");
+			btn.scale = vec2(1);
+			btn.color = rgb(1, 1, 1);
 		}
 	});
 
 }
 
-addButton("start", vec2(100, 100), () => debug.log("oh hi"));
-addButton("quit", vec2(100, 150), () => debug.log("bye"));
+addButton("start", vec2(200, 100), () => debug.log("oh hi"));
+addButton("quit", vec2(200, 200), () => debug.log("bye"));