+++ examples/jump.js
@@ -1,0 +1,128 @@
+const FLOOR_HEIGHT = 48;
+const JUMP_FORCE = 800;
+const SPEED = 480;
+
+// initialize context
+kaboom({ debug: true });
+
+// load assets
+loadSprite("bean", "sprites/bean.png");
+
+scene("game", () => {
+
+	// define gravity
+	gravity(2400);
+
+	addSky();
+
+	// add a game object to screen
+	const player = add([
+		// list of components
+		sprite("bean"),
+		pos(80, 40),
+		area(),
+		body(),
+	]);
+
+	// floor
+	add([
+		rect(width(), FLOOR_HEIGHT),
+		outline(4),
+		pos(0, height()),
+		origin("botleft"),
+		area(),
+		solid(),
+		color(0.45, 0.75, 1),
+	]);
+
+	function jump() {
+		if (player.grounded()) {
+			player.jump(JUMP_FORCE);
+		}
+	}
+
+	// jump when user press space
+	keyPress("space", jump);
+	mouseClick(jump);
+
+	function spawnTree() {
+
+		// add tree obj
+		add([
+			rect(48, rand(32, 96)),
+			area(),
+			outline(4),
+			pos(width(), height() - FLOOR_HEIGHT),
+			origin("botleft"),
+			color(0.95, 0.55, 1),
+			"tree",
+		]);
+
+		// wait a random amount of time to spawn next tree
+		wait(rand(0.5, 1.5), spawnTree);
+
+	}
+
+	// start spawning trees
+	spawnTree();
+
+	// define action for every game obj with tag "tree"
+	action("tree", (tree) => {
+
+		// all trees will move by 'SPEED' pixels per second every frame
+		tree.move(-SPEED, 0);
+
+		// if tree goes out of screen, we remove it to save performance
+		if (tree.pos.x <= -tree.width) {
+			destroy(tree);
+		}
+
+	});
+
+	// lose if player collides with any game obj with tag "tree"
+	player.collides("tree", () => {
+		// go to "lose" scene and pass the score
+		go("lose", score);
+		burp();
+	});
+
+	// keep track of score
+	let score = 0;
+
+	const scoreLabel = add([
+		text(score),
+		pos(24, 24),
+	]);
+
+	// increment score every frame
+	action(() => {
+		score++;
+		scoreLabel.text = score;
+	});
+
+});
+
+scene("lose", (score) => {
+
+	add([
+		sprite("bean"),
+		pos(width() / 2, height() / 2 - 108),
+		scale(3),
+		origin("center"),
+	]);
+
+	// display score
+	add([
+		text(score),
+		pos(width() / 2, height() / 2 + 108),
+		scale(3),
+		origin("center"),
+	]);
+
+	// go back to game with space is pressed
+	keyPress("space", () => go("game"));
+	mouseClick(() => go("game"));
+
+});
+
+go("game");