+++ examples/layer.js
@@ -1,9 +1,6 @@
-kaboom({
-	scale: 2,
-});
+kaboom();
 
-loadRoot("/pub/examples/");
-loadSprite("mark", "img/mark.png");
+loadSprite("bean", "sprites/bean.png");
 
 // layer "ui" will be on top of layer "game", with "game" layer being the default
 layers([
@@ -12,8 +9,8 @@ layers([
 ], "game");
 
 add([
-	sprite("mark"),
-	scale(10),
+	sprite("bean"),
+	scale(5),
 	// specify layer with layer() component
 	layer("ui"),
 	color(0, 0, 1),
@@ -21,7 +18,7 @@ add([
 
 // this obj doesn't have a layer() component, fallback on default "game" layer
 add([
-	sprite("mark"),
+	sprite("bean"),
 	pos(100, 100),
-	scale(10),
+	scale(5),
 ]);