+++ examples/text.js
@@ -1,36 +1,14 @@
-kaboom({
-	scale: 2,
-});
-
-loadRoot("/pub/examples/");
-// proggy is a default font
-loadProggy();
-// loadCGA();
-// load04b03();
-loadFont("04b03", "fonts/04b03_6x8.png", 6, 8);
-loadFont("CP437", "fonts/CP437_9x16.png", 9, 16, CP437_CHARS);
-
-const fonts = [
-	"unscii",
-	"04b03",
-	"proggy",
-	"CP437",
-];
-
-let curFont = 0;
+kaboom();
 
 const input = add([
-	text("123abc", 24, {
+	text("123abc", {
 		width: width(),
-		font: fonts[curFont],
 	}),
 ]);
 
 // listen to text input
 charInput((ch) => {
 	input.text += ch;
-	curFont = (curFont + 1) % fonts.length;
-	input.font = fonts[curFont];
 });
 
 // like keyPress() but will retrigger when key is being held (which is similar